+++ httpservice/response.go
@@ -1,0 +1,9 @@
+package httpservice
+
+import (
+	pwc "grab/week5/GrabGoTrainingWeek5Assignment/postwithcomment"
+)
+
+type PostWithCommentsResponse struct {
+	Posts []pwc.PostWithComments `json:"posts"`
+}
