--- renderer/renderer.go
@@ -1,5 +1,0 @@
-package renderer
-
-type Renderer interface {
-	Render(interface{}) ([]byte, string, error)
-}
